// frontend/src/api/schema.ts
export interface components {
    schemas: {
// ...
        Slot: {
            /** Format: date-time */
            startTime: string;
            /** Format: date-time */
            endTime: string;
            status: components["schemas"]["SlotStatus"];
        };
// ...
        SlotStatus: "available" | "booked" | "past" | "outside_window";
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}